"Freeing above" bugfix
Must not free nodes that are part of the chain! We still have to-free nodes
below, so checking the span does not work as a criterion. OTOH we can at this
point free everything that is not in the chain, because we have already moved
everything down that must be moved down first.