Bugfix: insertion into chain
Our complicated distinction of cases when inserting a node into the current
chain just got another update.

The main distinction is now whether we are inserting at the beginning of the
chain or elsewhere. In the former case, we cannot refer to chain[i - 1] because
i == 0. Also, we must account for the possibility that chain[i] has no parent.
On the other hand, we have the guarantee that chain[i] is not a leaf. In the
latter case, we can't call methods on chain[i] because it may be a leaf. On
the other hand, we can freely refer to chain[i - 1] and call methods on it.

BURP now runs without exception on our 14 real-world test cases! Next: check
whether the results actually make sense.